requirements.txt
fastapi
uvicorn
ollama
openai
python-multipart
python-dotenv
Pillow
chandra-ocr[hf]
mysql-connector-python
